refactor: 도메인기준 dao경로 변경, 사용하지 않는 set 메서드 제거
