+++ src/main/java/kitchenpos/orders/dao/OrderTableDao.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.orders.dao;
 
 import java.util.List;
 import java.util.Optional;
